pyScript/requirement.txt


flask
pyserial
flask-cors